Adds new file to implement position correction in python
This commit adds a new file to develop the position correction
functionality in python. Given the modular structure of the
package, the algorithm could be called as an engine after several
iterations of another engine. It could also be used to test new
arrangents and prototype modifications to current CUDA code.


Former-commit-id: 04ee0f551d8dda309a2f8da39275818f227214fa